utils.sh: fix the command to find cgroupv2 mount point
The commit 25886c1f921c ("utils.sh: fix find_controller_mount_point()
for cgroup v2") attempted to fix the logic to find the mount point of a
given controller in cgroupv2 setup but failed to remove the grep for the
controller (which is valid only for cgroup v1). This patch fixes it by
removing the extra grep.

Fixes: 25886c1f921c ("utils.sh: fix find_controller_mount_point() for cgroupv2)
Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>